Move RetryProperties class to within per-module properties
